Create routes for scrapbook pages
The routes are set up so that the following will work:

The first scrapbook in the list:
`/same/as/files/under/pages`
`/pages`
`/pages/new`
`/pages/same/as/files/under/pages`
`/pages/same/as/files/under/pages/edit`

If multiple scrapbooks:
`/:book/same/as/files/under/pages`
`/:book/pages`
`/:book/pages/new`
`/:book/pages/same/as/files/under/pages`
`/:book/pages/same/as/files/under/pages/edit`

The idea is that full URLs will start with the basename of the book and
the literal word "pages" as the path prefix, but we also allow for some
short versions of the path when viewing pages. (All of them are handled
by 'pages#show').