Isolate viewing the template pages in an iframe
Use an iframe for templates to make sure that there's no leaking of CSS
or JS from the Scrapbook library into the scrapbook pages users create.
This uses the concept of the "raw" path that already exists for files
that aren't HTML templates to help us separate rendering the Scrapbook
chrome and folder listing from rendering the template itself.