Add CI with GitHub Actions
I also had to make some changes to the Rubocop configuration to get it
to pass via GitHub Actions. I think it may have been installing a more
up-to-date version of the gem then I have locally installed. For some
reason this "update" counts method calls broken up on multiple lines as
more than 1 line of code and also includes the lines in the "ensure"
block as part of the line count for an RSpec example. I just bumped the
max up a couple to handle this regression.