Bump to version 0.2.0 in order to publish
There used to be a gem named Scrapbook, but it was yanked. In order to
publish a new Scrapbook gem, I need to use a version that wasn't already
published. At this time, it looks like the old gem published the
following versions:

0.1.2.1 - March 06, 2012 (6 KB) yanked
0.1.2 - March 06, 2012 (5.5 KB) yanked
0.1.1 - March 04, 2012 (5.5 KB) yanked
0.1 - March 04, 2012 (5.5 KB) yanked
0.0.1 - January 16, 2012 (3 KB) yanked

(Source: https://rubygems.org/gems/scrapbook/versions)

So I'm publishing this first release as 0.2