Enable Rails 6.1 support
There wasn't any reason to have the "propshaft" gem in the Gemfile since
the host application manages assets. (It was never required for the gem,
and having it in the Gemfile prevented the test suite from running for
Rails 6.) This removes that and makes slight modifications to the code
to allow Scrapbook to run appropriately in Rails 6 with Sprockets.