Fix empty directory listings
When rendering the root directory message or the directory message for
other folders without a corresponding template file, we were rendering
those templates directly instead of through the `iframe` to the `raw`
route. This fixes that so that those template files will get the same
styling as all the other files.

Along the way, this also cleans up and simplifies the `PagesController`
in light of the new architecture. We no longer need the `index` action
as everything can be sent to the `show` action while the `raw` action
does the heavy lifting.