Deprecate `scrapbook` and `pathname` locals in Scrapbook pages
These locals will eventually be removed in favor of using the `sb`
helper method. Code needs to be updated to call `sb.scrapbook` and
`sb.pathname` instead. This should mean that the only possible conflict
that the Scrapbook gem could have with the host's helper methods will
eventually just be `sb`.