Create helper to render Scrapbook page with its source preceding it
Created the `render_with_source` helper method for Scrapbook pages which
renders the source of the specified file, and then renders the page
itself.

This is great for sample code. You can create a folder named ".examples"
(the leading period will keep it hidden from the tree), and put sample
code in it. Your pages can then easily render the sample source code and
how it would render.